Don't upload any '*.orig' files
The current configuration prevents files created by git merge conflicts (ending in "*.orig") from being uploaded to Kibo, but only if they are in the root directory of the theme. This change excludes all "*.orig" files anywhere.

This is important especially for template files, because Hypr will load a "*.hypr.orig" file over a "*.hypr" file, displaying the conflict markers on the site.